Split browser_action index for readability
